Rename coordimage module
