+++ nilearn/_coordimage/coordimage.py
@@ -2,7 +2,7 @@ import nibabel as nib
 import numpy as np
 from nibabel.fileslice import fill_slicer
 
-import nilearn.coordimage.pointset as ps
+import nilearn._coordimage.pointset as ps
 
 
 class CoordinateImage:
