+++ nilearn/_coordimage/pointset.py
