+++ nilearn/_coordimage/tests/data/cifti2/ones.dscalar.nii
