+++ nilearn/_coordimage/tests/data/fsLR.wb.spec
