+++ nilearn/_coordimage/tests/data/fsaverage/surf/lh.white
